style(code): 패키지 및 클래스 취합
최종희
