+++ src/main/java/com/example/dicombridge/config/RestTemplateConfig.java
@@ -1,11 +1,11 @@
-package com.example.dicombridge.controller;
+package com.example.dicombridge.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
 
 @Configuration
-public class config {
+public class RestTemplateConfig {
 
     @Bean
     public RestTemplate restTemplate() {
