--- src/main/java/com/example/dicombridge/domain/StudyRepository.java
@@ -1,7 +1,0 @@
-package com.example.dicombridge.domain;
-
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StudyRepository extends JpaRepository<Study, Integer> {
-}
