--- src/main/java/com/example/dicombridge/domain/study/Imagetab.java
@@ -1,20 +1,0 @@
-package com.example.dicombridge.domain.study;//package com.example.dicombridge.domain.study;
-//
-//
-//import lombok.Getter;
-//import lombok.NoArgsConstructor;
-//
-//import javax.persistence.Entity;
-//import javax.persistence.Id;
-//import javax.persistence.Table;
-//
-//@Getter
-//@NoArgsConstructor
-//@Table(name = "imagetab")
-//@Entity
-//public class Imagetab {
-//    @Id
-//    private int studykey, serieskey, imagekey;
-//    private String studyinsuid, seriesinsuid, sopinstanceuid, sopclassuid, ststorageid,
-//
-//}
