+++ src/main/java/com/example/dicombridge/domain/study/StudyRequestDto.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class StudytabRequestDto {
+public class StudyRequestDto {
     private int studykey;
     private String studyinsuid, studydate, accessnum,  studyid, studydesc, modality;
     private String bodypart, pid,  pname, psex, pbirthdatetime, patage;
